Abstractization of the fixed funcs by making struct with changeable values

+++ VAR/src/Renderer/Utilities/CustomTypes/QueueFamiliyIndices.h
@@ -3,8 +3,8 @@
 
 struct QueueFamilyIndices
 {
-	uint32_t graphicsFamily = -1;
-	uint32_t presentationFamily = -1;
+	uint32_t graphicsFamily = (uint32_t)-1;
+	uint32_t presentationFamily = (uint32_t)-1;
 	
 
 	bool isComplete()
